Used `dynwave` to play audio and improve audio quality a lot
Now we don't have any popping, by adjusting the clock to a good tested frequency and using `dynwave` which will help reduce further

+++ trapezoid-core/src/lib.rs
@@ -279,7 +279,7 @@ impl Psx {
             .sync_gpu_and_blit_to_front(dest_image, full_vram, in_future)
     }
 
-    pub fn take_audio_buffer(&mut self) -> Vec<i16> {
+    pub fn take_audio_buffer(&mut self) -> Vec<f32> {
         self.bus.spu_mut().take_audio_buffer()
     }
